+++ www/modules/email/MessagesGrid.js
@@ -19,7 +19,7 @@ GO.email.MessagesGrid = function(config){
 	config.paging=true;
 
 	config.hideMode='offsets';
-		
+
 		config.cm =  new Ext.grid.ColumnModel({
 		defaults:{
 			sortable:true,
@@ -204,13 +204,7 @@ GO.email.MessagesGrid = function(config){
 				iconCls: 'ic-more',
 				text: t("Advanced", "email"),
 				handler: function(){
-					// var first = !this.searchDialog.dialog;
 					this.searchDialog.show();
-					// if(first) {
-					// 	this.searchDialog.dialog.on('hide', function() {
-					// 		this.searchField.updateView();
-					// 	}, this);
-					// }
 				},
 				scope: this
 			}]
@@ -258,24 +252,16 @@ GO.email.MessagesGrid = function(config){
 				this.emailClient.showAccountsDialog();
 			},
 			scope: this
-		}
-//		,{
-//			iconCls:'ic-view-compact',
-//			text: t("Toggle message window position", "email"),
-//			handler: function(){
-//				this.emailClient.moveGrid();
-//			},
-//			scope: this
-//		}
-,
+		},
 		this.showFlaggedButton
 		]
 	});
 
 
-	if(!config.hideSearch)
+	if(!config.hideSearch) {
 		config.tbar = [];
-	
+	}
+
 	GO.email.MessagesGrid.superclass.constructor.call(this, config);
 
 	var me = this;
@@ -415,6 +401,7 @@ GO.email.MessagesGrid = function(config){
 
 	//stop/start drag and drop when store loads when account is readOnly
 	this.store.on('load', function(store, records, options) {
+		console.log(store.baseParams);
 		if(this.getView().dragZone){
 			if(store.reader.jsonData.permission_level <= GO.permissionLevels.read) {
 				this.getView().dragZone.lock();