+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+28-03-2024: 6.8.40
+- Core: Added auth and lost password logs for fail2ban
+- Core: Made change password play nice with password managers
+- Core: Fixed upgrade SQL for MySQL 8
+
 25-03-2024: 6.8.39
 - Support / Core: GOUI was incompatible with 6.8 due to changes for 6.9+.
 - Core: prevent timing attack on password recovery
@@ -329,6 +334,8 @@
 - Core: Fixed bug in XSS detection
 - Core: Disallow modification of modifiedAt, createdAt, modifiedBy and createdBy via API.
 - Core: Create permissions were not checked on import
+- Email: make drafts editable in a mobile screen
+- Core: Added auth and lost password logs for fail2ban
 
 18-03-2023: 6.7.74
 - Core: fix searching for partial email addresses in global search